updates for AC4 and AC7
- move the call to getCalculatedBenefits to selectors/calculator.js and call the method getSchoolLocationsCalculatedBenefits
- update to correct logic for using the rates
